Simplify how the renderer is called
We don't need to explicitly pass the field object to the
renderer, since it is available in the agument hash.

We remove the Catalyst context from the argument list,
since that may make caching problematic.

We also only pass the field object at the very end since it
would be difficult to get a consistent signature for a field,
as it may include the entire form state.